+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This change log adheres to [keepachangelog.com](http://keepachangelog.com).
 
 ### Changed
 - Don't hide dropdown on blur event by default.
+- Don't activate the first dropdown item by default.
 
 ## [0.2.0] - 2016-02-29
 ### Added